iio: ad9517: Rework input clock error handling
The ad9517 can either use the CLK input or the REFIN clock in combination
with the VCO or both or generate the output clocks. This means which clocks
are required for proper operation depends on the configuration of the
device.

Currently the driver treats the REFIN clock as an optional clock but still
generates a warning if it is missing and requires the CLK clock as a
non-optional clock.

Re-work this to make both clocks optional and only generate a error message
if the configuration tries to use a clock that was not provided.

Furthermore only use devm_clk_get() to prevent leaking the clock handles.
And also make sure to only assume that no clock was provided when the error
code is -ENOENT rather than for every error code.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>