iio: axi_adxcvr: Remove unnecessary clock is_enabled() callback
The clock is_enabled() callback is primarily intended to be used to detect
clocks that are enabled during system boot, but are no longer needed. As
such the is_enabled() callback is supposed to report the state of the clock
read back from the hardware.

For the axi_adxcvr the clock is always disabled at system boot and the
current implementation of the is_enabled() callback just returns cached
software state. This is not necessary since the clock framework itself also
tracks the state in software and falls back to this if the is_enabled()
callback is not provided. This means the is_enabled() callback is redundant
and can be removed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>