drm/bridge: adv7511: Add support for configuring the EDID I2C address
The ADV7511 has a secondary I2C address for the EDID memory. This address
is programmable and currently the driver sets it to an address at a fixed
offset to the primary I2C address.

This approach does not work on all platforms and can lead to an I2C address
conflict with other devices. To avoid this allow the secondary address to
be specified through firmware data (e.g. devicetree).

This is done by replacing i2c_new_dummy() with i2c_new_secondary_device().
It will check if a address for the secondary I2C interface is specified via
the firmware interface and if it is use that address. If no address is
specified it will fallback to the fixed offset address.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>