iio: ad9467: ad9680: Handle optional SYSREF clock
The SYSREF clock is optional and might be omitted in the hardware design.
Handle the case where the SYSREF clock is not specified by setting the
sysref_clk pointer to NULL. All the clk API function calls are capable of
handling passing NULL as the clock, in this case they will turn in to
no-ops.

One thing that needs to be considered is that in subclass 1 the AD9680
waits for at least one SYSREF clock edge has been capture and the LMFC is
properly aligned before transitioning the JESD204 link to the DATA phase.

If the SYSREF signal is not supplied that means the converter will never
transition to the DATA phase. To avoid this use subclass 0 in this case.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>