iio: ad9361: Fix tuning diagnostics output
Newer kernels require that a continued printk statement is prefixed with
KERN_CONT, otherwise a newline will be inserted. Use the pr_info() and
pr_cont() macros to print the diagnostic output to make sure that the
proper message level indicators are used.

Also get rid of the extra newline at the end of the output, we don't really
need this.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>