iio: adc: ad9361: move register definitions to separate header file
The ad9361.h is a pretty big header file.
The intent is to be able to share some things amongst driver files without
exposing too much of each file's state, to prevent access to structs that
should only be controlled in specific files and nowhere else.

Specifically here, the aim is to make more some driver structs in
"ad9361.h" private, but the reg definitions are fine to be shared amongst
files.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>