iio: adc: ad9361: add ad9361_write_bist_reg() accessor to cache bist reg value
The value of the BIST_CONFIG register should also be stored on the cached
bist_config after it is written to the physical register.
Otherwise there could be a discrepancy between the cached value and the
value in the register.

To do this, the ad9361_write_bist_reg() accessor has been added, which
always caches the last written value.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>