iio: ad9517: Add AD9516 support
The AD9516 is mostly register map compatible to the ad9517. The only
difference is that it features two additional LVPECL outputs.

The AD9516 has different versions ranging from AD9516-0 to AD9516-5. These
different versions only differ in the supported locking range of the VCO.

This means the probe function in addition to the part id needs to know the
part family of a device. Encode both of them in the driver_data field of
the spi_device_id table.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>