iio: axi_adxcvr: Add support for when the ref clock is equal to the div40 clock
Currently the axi_adxcvr driver adjusts the div40 clock rate after the lane
rate has been updated.

But when the div40 clock is equal to the reference clock of the transceiver
PLL, like it is in some designs, it is necessary to adjust the reference
clock rate to the new target rate before computing the PLL parameters.

This enables dynamic reconfiguration of the lane rate on these platforms.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>