iio: ad9144: Add support for AD9135/AD9136
The AD9135 and AD9136 are a 11-bit and 16-bit respectively dual DAC up to
2.8 GSPS.

They are similar to the AD9144, the main difference being that they only
have two instead of four DAC channels. They share the same product ID as
the AD9144, but can be distinguished by the product grade.

Some of the JESD204 link modes are different and need special consideration
in the setup code. Otherwise device setup is the same.

Since for this part there are no legacy considerations use the 4 wire SPI
mode by default as it is intended by the Linux SPI framework.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>