iio: ad9508: Fix output channel phase calculation
The current output channel phase calculation does not consider that the
divider value that is stored in the register is the actual divider value
minus 1.

This can even result in a division by zero when trying to set the phase of
a output which has the divider set to 1.

This results in an incorrect phase display and phase update. Fix this by
adding 1 to the register divider value.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>